Pin api-go in order to resolve protojson.Marshal bug
This was fixed by https://github.com/temporalio/temporal/pull/5396 but
because there are incompatible changes between api-go, our sdk, and the
temporal server itself we haven't pulled this change into the server.

Those should be resolved soon but in the meantime we're going to pin the
version of the API used.