Adding namespace commands to cli (#476)
## What was changed
Adding operator namespace commands to cli. These are the commands that
are added
* operator namespace create
* operator namespace delete
* operator namespace describe
* operator namespace list
* operator nemspace update

## Breaking changes
* Removed --verbose option from `namespace update` command. This option
is used to print update request payload.
* Previous cli was printing historyArchivalUri and visibilityArchivalUri
fields only if the option `--fields` is set to `long` when executing
`namespace describe` command. Here we are always printing those fields
since we dont have `--fields` option.

## Why?

## Checklist

2. How was this tested:
Unit tests.

3. Any docs updates needed?