CLI Refresh: `workflow delete` command (#496)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Add support for workflow delete command. The functionality is mostly
unchanged from current CLI, but the output may be different.

## Why?
For the CLI refresh

## Checklist
<!--- add/delete as needed --->

1. How was this tested:
Tested in commands.workflow_test.go and via CLI with local dev
environment.

---------

Co-authored-by: Chad Retz <chad@temporal.io>