Batch rate limit (#551)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Added support for passing `--rps` flag to batch operations.

ℹ️ FYI I've only added a single test for TerminateWorkflow, even though
there are 3 batch APIs. Reason being that the API is the same and the
only difference is the "operation". Let me know if I should add it to
the others as well.

❓ Note sure why I'm seeing the CLA request on my PR.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested: Added new tests.

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->