Dump junit.xml (#650)
This PR switches the tests in CI to dump junit XML files. To do so, it
switches to use https://github.com/gotestyourself/gotestsum as the test
runner. (This is what the server repo uses also). An example of a build
with test failures is here:
https://github.com/temporalio/cli/actions/runs/10631569269/job/29472736222

As in https://github.com/temporalio/sdk-python/pull/617, the motivation
for this is that, given appropriate tools, it will make it easier for us
to study patterns of flakiness, test timings etc, across matrix builds
within a run, and across runs.