Create new `string-enum[]` type (#677)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR!
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

<!-- Describe what has changed in this PR -->
Created a new `string-enum[]` type

<!-- Tell your future self why have you made these changes -->
Command option behavior already exists, there just wasn't a specific
type for this yet.

<!--- add/delete as needed --->

1. Closes #670

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->

---------

Co-authored-by: David Reiss <dnr@dnr.im>