Auto generate CLI docs for documentation repo (#685)
## What was changed
Auto generate all of the *.mdx files present in the current
documentation repo. index.mdx is not auto-generated

IMPORTANT: The generated files are not yet being consumed anywhere. A
future PR will auto-publish a PR into the docs repo whenever there is a
docs related change.

Added a field in the YML to indicate doc-specific information, like SEO
related keywords and descriptions.

Removed `cmd-options.cdx` and print each command option description
inline in each file instead.
https://github.com/temporalio/documentation/pull/3122 removes all
references to this file from the documentation side.

## Why?
This is in a larger effort to have the CLI be the source of truth for
docs, and have the documentation auto-ingest any CLI changes, keeping
all information in sync.

This is the first step to this goal, there are a number of follow up
tasks that I wanted to split up from this PR:
1. Diff the current information on the docs with the CLI, and unify into
a single master-description for each command.
2. Auto-publish PR to docs repo whenever there is a CLI doc change
1. Auto update sidebars.js and vercel.js whenever a new temporal command
is added.
3. Clean up the SEO keywords/tags - being handled by the documentation
team
4. Support markdown in the CLI help output in terminal #687

## Checklist
7. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->
For now, docs will not be updated. There will be a future PR that
auto-creates a PR into the documentation repo with any updates.

---------

Co-authored-by: David Reiss <dnr@dnr.im>