Fix register search attributes when starting server (#653)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->
Fix register search attributes when starting server

## Why?
<!-- Tell your future self why have you made these changes -->
Currently, search attributes are registered after the server is ready,
which can cause race condition with other processes waiting for the
server to become ready too. This PR fixes this by registering the search
attributes together with the namespaces.

Address https://github.com/temporalio/temporal/issues/6195

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->