Add batch activity unpause. (#747)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->
<!--- For ALL Contributors 👇 -->


## What was changed
<!-- Describe what has changed in this PR -->
Update "activity unpause" API to support batch operation.

## Why?
<!-- Tell your future self why have you made these changes -->
Part of the activity API work.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->
https://temporalio.atlassian.net/browse/OSS-3789

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->
By adding test.

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->
Not yet.