Fix for DescribeWorkflow (#773)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->
Check if extended info is not nil.

## Why?
<!-- Tell your future self why have you made these changes -->
https://github.com/temporalio/cli/pull/771
Code assumes that some extended info exists in proto. Which is not true
for older server versions..