fix: activity pause,unpause,reset accept activity type OR activity id (#824)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
`temporal activity {pause,unpause,reset}` commands take either an
`--activity-id` or `activity-type` argument but not both. If both are
provided the command will fail

## Why?
We want to fail when both are provided, the user doesn't need to provide
activity type and id as type can be inferred from type.

## Checklist
<!--- add/delete as needed --->

1. Closes no issue

2. How was this tested:
* new unit test showing failure

3. Any docs updates needed?
documentation updates will be needed, @drewhoskins-temporal or
@ychebotarev can share where and I will update