Reset subcommand to update version (#812)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Added reset subcommands so that we can do reset + update version info
atomically in one operation. The server changes were made in
https://github.com/temporalio/temporal/pull/7719. This is the CLI
counterpart.

To allow subcommands we added a new flag `subcommands-optional`. When
this flag is present a parent command can be run in isolation without
any subcommands (i.e subcommands will be optional)

## Why?
So that we can reset + update_version at the same time.

## Checklist
<!--- add/delete as needed --->

1. Closes: <!-- add issue number here --> N/A

2. How was this tested:
Added tests + manually tested

3. Any docs updates needed?
N/A