display and update namespace replication state (#844)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Allow display and update namespace replication state.

## Why?
We want to check namespace replication state quickly from cli, and
update it if needed.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
Tested locally with server change
https://github.com/temporalio/temporal/pull/8234

3. Any docs updates needed?
(not sure if update is needed)
https://docs.temporal.io/cli/operator#update