Add `temporal worker deployment manager-identity` commands (#858)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Add `temporal worker deployment manager-identity` commands (`set` and
`unset`)

## Why?
So that CLI clients of Worker Deployments that are managed by an
automated system (such as the Kubernetes Worker Controller) can safely
coordinate with that client to make intended changes and ensure they are
not overwritten.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
Added a new test. Also tested in SDK and server.

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->