Move to internal package (#874)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

**Before**

```
/temporalcli
  /internal
     /cmd
     /...
  commands.go
```

**After**

```
/internal
  /cmd
  /temporalcli
     /...
     commands.go
```

Also see the related
https://github.com/temporalio/documentation/pull/3997.

## Why?
<!-- Tell your future self why have you made these changes -->

We want to establish a `cliext` that will contain the shared code for
[writing CLI
extensions](https://github.com/temporalio/proposals/blob/master/cli/cli-extensions.md).

> The existing github.com/temporalio/cli/temporalcli package should not
be used and may be moved to internal as part of this project.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->