Commands generator binary (#875)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Make `gen-commands` and `gen-docs` a binary that can be used any YAML
input.

## Why?
<!-- Tell your future self why have you made these changes -->

Temporal wants to re-use the `commandsgen` tool for other _internal_
CLIs.

**It is not supported for users outside of Temporal.** 

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->