Enable setting of Temporal headers through cli (#883)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Support setting of Temporal headers in workflows

## Why?
The cli doesn't support the setting of workflow headers. With this
change we would be able to pass the headers to Workflows using cli.

## Checklist
<!--- add/delete as needed --->

1. Closes - https://github.com/temporalio/cli/issues/876

2. How was this tested: 
- Brought up the dev server using : `temporal server start-dev`
- Build the cli locally and passed the `--headers` flag to start a
workflow. Checked in Temporal UI, the Header section shows the headers
set in Workflow.

3. Any docs updates needed?
https://docs.temporal.io/cli/workflow