Move gen commands (#887)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Moved `‎internal/cmd/gen-commands` to `cmd/gen-commands`; and same for
`gen-docs`.

## Why?
<!-- Tell your future self why have you made these changes -->

When they are in `internal/` they cannot be invoked via `go run` from
outside the repository; making it cumbersome to use.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->