Add pause/unpause commands (#885)
Redoing https://github.com/temporalio/cli/pull/885/
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR!
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

Added 2 new commands (experimental)
- temporal workflow pause
- temporal workflow unpause

Did a bunch of unrelated cleanup as well.

Need these commands to allow pausing and unpausing of workflows via cli

<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->