Invoke CLI extension (#889)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

CLI now invokes extensions in accordance with [CLI Extensions
proposal](https://github.com/temporalio/proposals/blob/master/cli/cli-extensions.md).

Note that the help command integration part is going to be in a
follow-up PR.

## Why?

Allow Temporal CLI Extensions.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->